ec1 functions mostly work on ec2 hardware
What works:

EC1 Functionality is mostly a reality.
- Chassis boots and inits variables and does system check
- Reacts properly to door latch (when it’s closed, the disinfection
state machine can run, when it’s open, the buzzer beeps and you can’t
start a disinfection cycle)
- Status prints on debug port
- Blue scrolling display mostly displays what it’s supposed do during
Ready, Clean, Cleaning, Dirty, etc.
- Fake keypad input from the debug port: type ‘k’ or ‘K’ and the chassis
will scan for devices and kick off the solenoid
- Device scan/detection kinda works, thresholds need to be tweaked,
picks up big devices but not little ones
- LED board tracking numbers increment, at the right rate? Not sure.
- LED board serial numbers detected and reported properly
- LED and power supply control appear to work properly

What’s not working (Julia):
- Need to modify disinfection state machine for proper keypad function.
For EC2, keypad should allow you open the door. Right now I am using the
fake keypad input from the debug port as an action button from the old
chassis which is not right.
- Shelves with devices are identified on the blue scrolling display, but
it’s supposed to be for 8 seconds per shelf and it’s going faster than
that, need to fix that.
- Not storing usage data in flash, so when you cycle power on the
chassis, usage data disappears (won’t change this for the trade show,
need lots of testing on flash access code before you send it to a show,
don’t want it to hang)
- Tweak device scan/detection thresholds
- Debug output doesn’t properly echo characters back to the user (surely
dumb)
- Not able to do floating point math for the disinfection time
calculation like on the UC3C for some reason
- Disinfection cycle stays on perpetually, I think this was from the
code modified for the microbiologists, we will adjust based on whatever
is needed for the trade show