microbiologists
Same thing we did in January '16 with the GEN 1 hardware. The
microbiologists want to be able to run this thing ad infinitum so they
can determine effectiveness. But I have a cpu_timeout() mechanism
limitation, the longest thing it will time is maybe 2-3 minutes, it
doesn't work for 30 minutes. So I changed this code to just not turn the
shelves off, just leave them on unless there is a chassis error or the
door opens. Worked in January. Hopefully ok this time.